ISS_Tracker/requirements.txt
azure-functions
requests>=2.31,<3
tzdata>=2024.1 ; sys_platform == "win32"